+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 ====================
- Unpublished
+ 5.3.0 - 2022-02-17
 ====================
 
 Added